requirements.txt

google-cloud-texttospeech==0.5.0
xlrd==1.2.0
pydub==0.23.1
pathlib==1.0.1
argparse==1.4.0

